+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Experimental feature warning for `x402 init` and `x402 import` commands
+
 ## [0.2.3] - 2026-03-31
 
 ## [0.2.2] - 2026-03-31